Allow a more general reference solution.
This solution allows a range of patterns to be used as report widgets,
and makes the interface to the Report Widgets smaller.

This restricts the contract to the report widgets to just the two
interactions (construction and call to #render).

It also allows widgets to be referenced by any javascript expression,
such as the name of a class or a call to a function (such as a factory
function or a require call).